merge duplicated NeighborStat.get_stat
Note: this is a simple fix to resolve #1088, but I think we should design a clear architecture to call neighbor stat.

This should reduce the half of the time, but it may be still too long. We can consider some better algorithm to calculate neighbour stat (like KDtree?) for further optimization.