add a graph optimizer to parallelize prod_force_a_cpu
This commit adds a custom graph optimizer to enable the parallelization for prod_force_a_cpu. The optimizer is on-the-fly and will not affect saved graph, so one can continue training or inference without the optimizer enable.
The OP is splited into multiple OPs, which are then added together, so it can be run across multiple steams, by controling TF_INTER_OP_PARALLELISM_THREADS. Without optimizer, there's only one steaming working and others are resting. A performance improvement of this OP can be obversed in the water example and also in my own case.
The optimizer is disabled by default and the new attribute of the OP has a default value, so there's no breaking change. One can enable it by environment variables.
A UT is added to confirm the result is not changed.