uncouple model precision and interface precision
This commit allows loading float models using double interface, or the other way around.
The model precision is perceived from the type of `descrpt_attr/rcut`. All `run_model` and `session_input_tensors` are rewritten as templates.
This is an important step of #1948. Next functions in the whole module can be migrated to templates.