skip `get_gpus` subprocess when TF is cpu only (#2135)
When I benchmark deepmd-kit on my machine, I found `get_gpus` takes
about 2s and is quite slow. In #905, a subprocess was added to get
available GPUs. As benchmarked in #2121, it's quite slow to import
tensorflow.

I don't have better ideas not to call a subprocess, but we can skip this
process when TensorFlow is not built against GPUs. The tests on the
GitHub Actions will also benefit.

Attached is the selected profiling:
>  ncalls  tottime  percall  cumtime  percall filename:lineno(function)
> 1    0.000    0.000    2.141    2.141 local.py:15(get_gpus)
> 1    0.000    0.000    2.133    2.133 subprocess.py:1090(communicate)

Signed-off-by: Jinzhe Zeng <jinzhe.zeng@rutgers.edu>