avoid unnecessary std::endl
std::endl = \n + flush, and in these situations, it is not required to flush every line:
(1) multiple lines into stdout and stderr. Only the last line is required to flush;
(2) frequently writing to a file. In the deepmd-kit, model_devi.out was flushed every line but usually one will not monitor it in real time.

See also:
https://en.cppreference.com/w/cpp/io/manip/endl
https://stackoverflow.com/a/17468264/9567349

Signed-off-by: Jinzhe Zeng <jinzhe.zeng@rutgers.edu>