Support different learning rate settings for each fitting net in multi-task mode (#2348)
It is a breaking change to the user interface in multi-task mode. Please
review carefully.

Replace "learning_rate" with "learning_rate_dict" in multi-task mode to
support different learning rate settings for each fitting net, which is
familiar to "loss_dict". The format of `lcurve.out` will be like the
following:
```
#  step   sys1_rmse_val sys1_rmse_trn   sys1_rmse_e_val sys1_rmse_e_trn   sys1_rmse_f_val sys1_rmse_f_trn   sys1_lr  
 sys2_rmse_val sys2_rmse_trn   sys2_rmse_e_val sys2_rmse_e_trn   sys2_rmse_f_val sys2_rmse_f_trn   sys2_lr
...
```
The learning rate of each system will be printed separately.

---------

Signed-off-by: HuangJiameng <105633685+HuangJiameng@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>