fix dplr: correct type check in get_valid_pairs (#2580)
The binary_search method in C++ is only valid for sorted vectors. 
The following case in "type_associate" of "fix dplr" will raise error: 4
5 3 6. 4 and 3 are real atom types and stored in sel_type. The vector
sel_type is sorted. 5 and 6 are associated with 4 and 3, respectively.
The resulting sel_type vector is 3 4, and the resulting dpl_type vector
is 6 5. If a pair 4 5 is searched, then this line



https://github.com/deepmodeling/deepmd-kit/blob/e7d68c165c20aea04af686f8a81b179154b1d679/source/lmp/fix_dplr.cpp#L195-L196

will do binary search for 4 in 3 4, which results in True; and will
search for 5 in 6 5, which results in False.

Therefore, I am copying dpl_type vector to the dpl_type_sorted vector
and sort it. The binary search should be used on the sorted vector.

## The above solution is abandoned.
In stead, do not use binary_search in the type check of get_valid_pairs.
The current method is: first find the type of the real atom, and check
if the type of its bonded atom is consistent with that defined by
type_associate.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>