fix segfault in ~Region (#3108)
When debuging #3103, I notice a segfault in `~Region`, preventing the
actual error message thrown. `_norm_copy_coord_gpu` replaces `boxt` and
`rec_boxt` of `Region` with GPU pointers, runs `deepmd::copy_coord_gpu`,
and finnally recover the original pointer that can be deleted. However,
`deepmd::copy_coord_gpu` might throw CUDA errors for any reason, and the
pointers are not recovered. `~Region` tries to delete a pointer that it
doesn't own, causing the segfault. The CUDA error message is not visible
due to segfault. The segfault in #2895 may be also caused by it.
This PR adds a new constructor to `Region` to accept the external
pointers. `~Region` will delete `boxt` and`rec_boxt` only when the
pointer is not external.
We still need to figure out the reason for the error of `copy_coord_gpu`
behind the segfault.

---------

Signed-off-by: Jinzhe Zeng <jinzhe.zeng@rutgers.edu>