PT: keep the same checkpoint behavior as TF
Set the default save_ckpt to `model.ckpt` as the prefix. When saving checkpoints, `model.ckpt-100.pt` will be saved, and `model.ckpt.pt` will be symlinked to `model.ckpt-100.pt`. A `checkpoint` file will be saved to record `model.ckpt-100.pt`.

This keeps the same behavior as the TF backend.

Signed-off-by: Jinzhe Zeng <jinzhe.zeng@rutgers.edu>