c: change the required shape of electric field to nloc * 3
A segfault sometimes appears in the tests after #3223. The reason is that the required shape of electric field is set to nall * 3 in the C interface, but a vector of nloc * 3 is actually given in the tests and lammps and used in the C++ interface. It's only found by sanitizer as the program doesn't write to the invalid address (only reading it won't cause segfault).

Signed-off-by: Jinzhe Zeng <jinzhe.zeng@rutgers.edu>