c: change the required shape of electric field to nloc * 3 (#3237)
[A
segfault](https://github.com/deepmodeling/deepmd-kit/actions/runs/7782245372/job/21218255452)
sometimes appears in the tests after #3223. The reason is that the
required shape of the electric field is set to nall * 3 in the C
interface, but a vector of nloc * 3 is given in the tests and lammps and
used in the C++ interface. It was not caught before, as the program
didn't write to these addresses (only reading it usually won't cause a
segfault, unless the address is invalid).
Perhaps fix #2895.

---------

Signed-off-by: Jinzhe Zeng <jinzhe.zeng@rutgers.edu>