pt: avoid `set_default_dtype` in tests
Why it is bad:
The default dtype in the producution code is still float32. When setting it to float64 during tests, the actual producation behavior may not be properly tested. (for example, if the production code misses dtype for pt.zeros, we are not able to find it in these tests)

Signed-off-by: Jinzhe Zeng <jinzhe.zeng@rutgers.edu>