pt: explicitly set device
Force requires setting device to `env.DEVICE` or `cpu` explictly for functions like `torch.tensor`, `torch.zeros`, `torch.ones`, `torch.rand`, `torch.eye`, and dataloader. This ensures that no OPs that should be run on GPUs run on CPUs.
The trick here is `torch.set_default_device("cuda:9999999")` in the tests, so errors will be thrown if no device is set.

Tips:
(1) Avoid `torch.zeros(...).to(device=...)`. This firstly initlizes memory on CPUs and copy it to GPUs.
(2) Use `with torch.device(...)` for third-party modules.

Signed-off-by: Jinzhe Zeng <jinzhe.zeng@rutgers.edu>