pt: explicitly set device (#3307)
Forcely requires setting the device to `env.DEVICE` or `cpu` explicitly
for functions like `torch.tensor`, `torch.zeros`, `torch.ones`,
`torch.rand`, `torch.eye`, `LayerNorm`, and data loader.
This ensures that no OP runs on the wrong device. The trick here is
`torch.set_default_device("cuda:9999999")` in the tests, so errors will
be thrown if the default device is used.

Tips:
(1) Avoid `torch.zeros(...).to(device=...)`. This first initializes
memory on CPUs and copies it to GPUs.
(2) Use `with torch.device(...)` for a module that cannot set a device
(i.e., the data loader).

---------

Signed-off-by: Jinzhe Zeng <jinzhe.zeng@rutgers.edu>