allow loading either nsel or natoms atomic tensor data
A new parameter "output_natoms_for_type_sel" is added. (default=false)
If sel_types is given, output_natoms_for_type_sel is true, and the data dimesion is nsel, it will be converted to natoms.
If sel_types is given, output_natoms_for_type_sel is false, and the data dimesion is natoms, it will be converted to nsel.
In other situations, it keeps the original shape.

Signed-off-by: Jinzhe Zeng <jinzhe.zeng@rutgers.edu>