fix(tf): make `se_atten_v2` masking smooth when davg is not zero (#3632)
Currently, `se_atten_v2` is always masked to zero when `exclude_types`
is given. However, for the no neighbor case, the placeholder for a
virtual neighbor is `davg`. This causes discontinuity when
`set_davg_zero` is not set.

This PR uses `davg` for masking.

In production, we usually use `set_davg_zero` along with
`exclude_types`, so it hasn't caused a real problem.

I notice PT hasn't implemented `se_atten_v2` or `exclude_types`, but we
need attention in the future.

---------

Signed-off-by: Jinzhe Zeng <jinzhe.zeng@rutgers.edu>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>