ci(wheel): build PT OPs (#3894)
Build PT OPs for our wheels (against the CPU version of PT but work with
the GPU version). Add PT to the build dependencies if it is not found.
Bump MPICH to the latest version. Update documentation to address that
PT support for LAMMPS and i-PI is not included.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Introduced PyTorch support across various backend functionalities,
enhancing compatibility and capabilities alongside existing TensorFlow
support.

- **Documentation**
- Updated installation guide to include warnings about TensorFlow
backend support and removed outdated notes about macOS arm64 package
support.

- **Chores**
- Updated dependency versions and build configurations to include new
PyTorch settings and improved compatibility for macOS and Linux.
- Adjusted build scripts to conditionally link libraries based on new
PyTorch-related parameters.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Signed-off-by: Jinzhe Zeng <jinzhe.zeng@rutgers.edu>