feat: support array API
Fix #3430.
This PR sets up the basic support for the array API, and make an example function (`compute_smooth_weight`) to support the array API.
I believe NumPy and JAX have supported it (or through `array-api-compat`), so we don't need to write things twice for NumPy and JAX (although we can write them using the ChatGPT, it's still better to maintain only one thing). There are some challeging to use it in the TorchScript, so I give it up.
Supporting more function can be implemented in the following PRs.

Signed-off-by: Jinzhe Zeng <jinzhe.zeng@rutgers.edu>