fix(pt): make PT training step idx consistent with TF
Fix #4206.
Currently, the training step index displayed in TF and PT has different meanings:
- In TF, step 0 means no training; step 1 means a training step has been performed. The maximum training step is equal to the number of steps.
- In PT, step 0 means a training step has been performed. The maximum training step is the number of steps minus 1.
This PR corrects the defination of the step index in PT and makes them consistent.
There are still a difference: TF shows step 0 but PT shows step 1. Showing step 0 in PT needs heavy refactor and thus is not included in this PR.

Signed-off-by: Jinzhe Zeng <jinzhe.zeng@rutgers.edu>