Refactor property fitting interface (#4471)
Two main changes：
1. The program can read data files in `npy` format with any
prefix(`tc.npy`, `band_gap.npy`...). One just needs to write the name of
the property and the corresponding dimension in `model/fitting` in
`input.json`.
2. Data normalisation has been added to the program. Specifically, the
mean and standard deviation of the properties are calculated when
calculating the stat, the output of `fitting_net` is multiplied by the
standard deviation and then added to the mean to get the output of the
`PropertyModel`, and when calculating the loss, the loss is again
normalised.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

## Release Notes

- **New Features**
- Introduced new parameters for property loss calculations and model
fitting, enhancing flexibility and functionality.
- Added methods for retrieving property names and checking property
intensity across various classes.

- **Bug Fixes**
- Improved validation and error handling for property-related
calculations.

- **Documentation**
- Enhanced documentation for property fitting, including detailed
parameter explanations and structured training examples.
- Added documentation for new parameters in the JSON configuration
related to property fitting.

- **Tests**
- Added new test cases to validate the functionality of updated methods
and properties.
- Updated existing tests to utilize dynamic property names instead of
hardcoded strings.

- **Chores**
- Updated configuration files and test data to align with new property
handling features.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Signed-off-by: Chenqqian Zhang <100290172+Chengqian-Zhang@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>