fix(jax): set `default_matmul_precision` to `tensorfloat32` (#4726)
See https://github.com/jax-ml/jax/issues/24909. Without setting this
flag, the precision will become very low, which seems to be a bug (the
documentation says the GPU uses tensorfloat32 or float32, but the
default behavior seems wrong...).
See
https://docs.jax.dev/en/latest/_autosummary/jax.default_matmul_precision.html
for what this option is.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Chores**
- Updated environment configuration to set default matrix multiplication
precision to "tensorfloat32" for improved performance with JAX.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

Signed-off-by: Jinzhe Zeng <jinzhe.zeng@ustc.edu.cn>